requirements.txt
asgiref==3.6.0
Django==3.2.5
django-bulk-update-or-create==0.3.0
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
iniconfig==2.0.0
Markdown==3.4.3
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
pypng==0.20220715.0
PyQRCode==1.2.1
pytest==7.3.1
pytest-mock==3.10.0
pytz==2023.3
qrcode==7.4.2
rest-api-payload==0.0.7
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
whitenoise==6.4.0
